Change from constant block value being subtracted from damage on block to reducing the hit by 30% for targets.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@6551 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/simulationcraft.h
@@ -3395,7 +3395,7 @@ struct target_t
   int level;
   int spell_resistance[ SCHOOL_MAX ];
   int initial_armor, armor;
-  int block_value;
+  double block_value;
   double attack_speed, attack_damage, weapon_skill;
   double fixed_health, initial_health, current_health;
   double fixed_health_percentage;
